requirements.txt
flask==2.3.2
neo4j==5.24.0
python-dotenv==1.0.0
requests==2.31.0
spacy==3.7.2
